LayoutTests/platform/gtk/accessibility/table-hierarchy-expected.txt

foo	bar
Odd	Even
1	2
3	4
hello	world
This tests the Atk table hierarhy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS table1.role is 'table'
PASS table2.role is 'table'
PASS table3.role is 'table'
PASS children is 2
PASS table1.childAtIndex(i).role is 'table cell'
PASS table1.childAtIndex(i).role is 'table cell'
PASS children is 6
PASS table2.childAtIndex(i).role is 'table cell'
PASS table2.childAtIndex(i).role is 'table cell'
PASS table2.childAtIndex(i).role is 'table cell'
PASS table2.childAtIndex(i).role is 'table cell'
PASS table2.childAtIndex(i).role is 'table cell'
PASS table2.childAtIndex(i).role is 'table cell'
PASS children is 2
PASS table3.childAtIndex(i).role is 'table cell'
PASS table3.childAtIndex(i).role is 'table cell'
PASS successfullyParsed is true

TEST COMPLETE
